Started code rewrite/cleanup; added TODO

+++ backlightctl.c
@@ -2,43 +2,15 @@
 #include <stdio.h>
 #include <signal.h>
 #include <unistd.h>
-#include <errno.h>
-
 #include "common.h"
 #include "procparse.h"
 
-#define VERSION VERSS
-#define CONFIG_STRING CONFIGS
-
-static void usage(void);
-static void printbckl(void);
-static void backlightctl(char* const);
-
-int main(int argc,char ** argv)
-{
-	if(argc == 1)
-		usage();
-	if(argc > 1){
-		switch(argv[1][0]){
-			case 'p':
-				printbckl();
-				break;
-			case 's':
-				if(3 == argc)
-					backlightctl(argv[2]);
-				else
-					usage();
-				break;
-			default:
-				usage();
-				break;
-			}
-	}
-}
+/*#define VERSION */
+/*#define CONFIG_STRING*/
 
 static void usage(void)
 {
-	fprintf(stdout,"Usage:\n"\
+	fprintf(stdout, "Usage:\n"\
 			"backlightctl [OPTION] [VALUE]\n"\
 			"Where Option is one of:\n"\
 			"p:\n Print current backlight scale in percent\n"\
@@ -51,54 +23,58 @@ static void usage(void)
 
 static void printbckl(void)
 {
-	FILE* strm = fopen(CONFIG_STRING,"r");
-	char ch = 0;
-	char * buf = NULL;
-	int length = 0;
-	for(;(ch=fgetc(strm))!= EOF;length++){
-		if(ch == '\n' ) /* We don't need newlines */
-			break;
-		buf=addread(buf,ch,length);
-		}
-	fclose(strm);
-	buf=addread(buf,'\0',length);
+	FILE* strm = NULL;
+	char buf[4]; /* Not more than 3 chars unless you wanna set it to 1000%*/
+
+	if ((strm = fopen(CONFIG_STRING,"r")) == NULL)
+		die("Open");
+	if (fgets(buf, 3, strm) == NULL)
+		die("Fgets");
 	printf("Brightness scale: %s%%\n",buf);
-	free(buf);
+	fclose(strm);
 }
 
 static void backlightctl(char * const value)
 {
 	int scale = atoi(value);
-	if((scale < 10) ||(scale > 100)){
+	if ((scale < 10) ||(scale > 100)) {
 		fprintf(stdout, "Wrong value!\n");
 		usage();
-	}
-	else{
-		FILE* strm = fopen(CONFIG_STRING,"w");
+	} else {
+		FILE* strm = NULL;
+		if ((strm = fopen(CONFIG_STRING,"w")) == NULL)
+			die("Open");
 		fprintf(strm,"%d", scale);
 		fclose(strm);
 		pid_t backlightd_pid = procparse("backlightd");
-		if(backlightd_pid == 0)
-			fprintf(stderr,"Could'nt find pid,check for proc"\
+		if (backlightd_pid == 0)
+			fprintf(stderr,"Could'nt find pid, check for proc"\
 					"Permissions!");
-		if(kill(backlightd_pid,SIGUSR1)!= 0){
-#ifdef DEBUG
-			int err = errno;
-#endif
-			fprintf(stdout,"Call failed!\n");
-#ifdef DEBUG
-			switch(errno){
-				case EPERM:
-					puts("No perms!");
-					break;
-				case EINVAL:
-					puts("Invalid value");
-					break;
-				case ESRCH:
-					puts("Does not exist");
-					break;
+		if (kill(backlightd_pid, SIGUSR1)!= 0)
+			die("Kill");
+	}
+}
+
+int main(int argc,char ** argv)
+{
+	if (argc == 1)
+		usage();
+	if (argc > 1) {
+		switch(argv[1][0]) {
+		case 'p':
+			printbckl();
+			break;
+		case 's':
+			if (3 == argc) {
+				backlightctl(argv[2]);
 			}
-#endif
+			else {
+				usage();
+			}
+				break;
+		default:
+				usage();
+				break;
 		}
 	}
 }
